fix save absoluteLeasetime crash

+++ services/dhcp_client/include/dhcp_client_def.h
@@ -399,7 +399,7 @@ typedef struct{
 
 struct IpInfoCached {
     std::string bssid;
-    uint32_t absoluteLeasetime;
+    int64_t absoluteLeasetime;
     struct DhcpIpResult ipResult;
 };
