+++ services/dhcp_client/src/dhcp_result_store_manager.cpp
@@ -192,7 +192,7 @@ int32_t DhcpResultStoreManager::SetClassKeyValue(IpInfoCached &item, const std::
     if (key == "bssid") {
         item.bssid = value;
     } else if (key == "absoluteLeasetime") {
-        item.absoluteLeasetime = static_cast<uint32_t>(std::stol(value));
+        item.absoluteLeasetime = static_cast<int64_t>(std::stol(value));
     } else if (key == "strYiaddr") {
         if (strncpy_s(
             item.ipResult.strYiaddr, sizeof(item.ipResult.strYiaddr), value.c_str(), value.size()) != EOK) {
